requirements.txt
gsheetsdb==0.1.13.1
pandas==1.1.4
Pillow==9.2.0
pymongo==4.2.0
snowflake==0.0.3
SQLAlchemy==1.4.41
streamlit==1.12.0
plotly==5.8.2
snowflake-connector-python==2.7.9
snowflake-sqlalchemy==1.4.1